fnatic-trading-bots/requirements.txt
git+https://github.com/openai/swarm.git
python-dotenv
openai
cdp-sdk
google-generativeai
flask